Moved terms storage functionality to the original storage
